+++ History.md
@@ -1,3 +1,13 @@
+# 0.5.0 (2014-06-16)
+ * [new] add packages search bar
+ * [new] get buildlog preview by default instead of datestamp when click on a package
+ * [new] always show datestamp information below the title
+ * [new] add page Commands which documents out the tasks interface of Deb-o-Matic
+ * [new] better tooltip over the action buttons (such as Download, View al file, Get all)
+ * [new] show the debomatic architecture in the header description
+ * [new] show architecture and version in footer if header is hidden in preferences
+ * [fix] when watching a file which gets deletion tail lib crashes
+
 # 0.4.2 (2014-06-15)
  * [fix] calculate pidfile for debomatic status check after merging configurations
 
@@ -20,7 +30,7 @@
 
 # 0.2.3 (2014-06-06)
  * [new] get all sources in one click
- * [fix] code valdiatation via jshint
+ * [fix] code validatation via jshint
 
 # 0.2.2 (2014-06-04)
  * [fix] upgrade to express 4.x